Adiciona o código referente ao Trabalho 5
Exclusão das faces e inclusão de Aliens que possuem minions lacaios que
giram e atiram.

+++ src/Game.cpp
@@ -63,9 +63,9 @@ void Game::Run(){
 }
 
 void Game::CalculateDeltaTime(){
-    Uint32 newFrame = SDL_GetTicks();
+    float newFrame = SDL_GetTicks();
 
-    dt = (float) (newFrame - frameStart) / 1000;
+    dt = (newFrame - frameStart) / 1000;
     frameStart = (float) newFrame;
 }
